Refactor test harness
