+++ tests/README.md
@@ -3,3 +3,11 @@
 ## Unit Tests
 
 A battery of tests with the goal to ensure that any changes made will not hopefully pooch the entire project. These have a long way to go to test all of the things since we keep adding new toys to the interface.
+
+Legacy = Tests implemented before refactor, but have not yet been updated to the new testing framework
+
+Units = Tests of a specific class or module only
+Integration = Tests of how classes/modules work together
+System = Top level tests that simulate the user experience
+
+Resources = Files needed to help determine if the tests work (e.g. test images)
