+++ tests/legacy/legacy_waterfalls.py
@@ -3,7 +3,7 @@
 #
 #in general, these tests should be testing the import framework
 
-from test_common import *   #contains utilities common to all tests.  Should come first before the other imports.
+from tests.legacy.common_vars import *   #contains utilities common to all tests.  Should come first before the other imports.
 
 import unittest     #for unit testing, har har
 import sys			#so that we can check to see which modules are imported at any given time
